+++ packages/flagmoji/CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - 2024-05-16
+
+### Fixed
+
+- Vietnam naming fix from Viet Nam to Vietnam
+
+### ADDED
+
+- Country name slugs
+
 ## [0.1.4] - 2024-03-20
 
 ### Fixed